Add "$Q" key in header for page numbers that don't reset
abcm2ps supports a "$P" key in header or footer lines that will output
the current page number and can be reset using "%%newpage". In my
books of SCD ball sets, I like to include both the page number inside
the tune and the page number inside the book on every page, so the
"$Q" adds another page number that is not changeable by the user.